Prevent removing Safest profile
Track the garlic level that a profile is created with, if any, and
prevent removing Safest profiles.

Issue: calyxos#1518

+++ app/src/main/java/org/calyxos/bellis/ui/ManagedProfileFragment.kt
@@ -8,6 +8,7 @@ package org.calyxos.bellis.ui
 
 import android.app.Dialog
 import android.app.admin.DevicePolicyManager
+import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.os.Process
@@ -18,11 +19,19 @@ import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.Fragment
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import org.calyxos.bellis.R
+import org.calyxos.bellis.utils.PostProvisioningHelper
 
 class ManagedProfileFragment : Fragment(R.layout.managed_profile_fragment) {
 
     private val TAG = ManagedProfileFragment::class.java.simpleName
     private val userSettings = "android.settings.USER_SETTINGS"
+
+    companion object {
+        fun shouldAllowRemoveProfile(context: Context): Boolean {
+            return PostProvisioningHelper.getGarlicLevel(context) !=
+                PostProvisioningHelper.GarlicLevel.SAFEST
+        }
+    }
 
     class RemoveProfileDialogFragment : DialogFragment() {
 
@@ -40,7 +49,9 @@ class ManagedProfileFragment : Fragment(R.layout.managed_profile_fragment) {
                 .setTitle(getString(R.string.remove_profile))
                 .setMessage(getString(R.string.remove_profile_confirmation))
                 .setPositiveButton(android.R.string.ok) { _, _ ->
-                    context?.getSystemService(DevicePolicyManager::class.java)?.wipeData(0)
+                    if (shouldAllowRemoveProfile(requireContext())) {
+                        context?.getSystemService(DevicePolicyManager::class.java)?.wipeData(0)
+                    }
                 }
                 .setNegativeButton(android.R.string.cancel, null)
                 .create()
@@ -49,6 +60,13 @@ class ManagedProfileFragment : Fragment(R.layout.managed_profile_fragment) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        view.findViewById<View>(R.id.removeProfile)?.visibility =
+            if (shouldAllowRemoveProfile(view.context)) {
+                View.VISIBLE
+            } else {
+                View.GONE
+            }
 
         view.findViewById<Toolbar>(R.id.toolbar)?.setOnMenuItemClickListener {
             when (it.itemId) {
